Rever joining kv subcommands into one source file
I thought I would like having all of the kv subcommands
in one file. I was wrong. The commands are logically separated
anyway so there isn't really a benefit to having them together.
